Fix IPMI recon script so that it works with new style custom fields
Updates the IPMI recon script to pass parameters to the IPMI plugin using the
new style custom fields.

Add a field for specifying additonal options to ipmi-sensors.  This is required
for IPMI v2.0 devices and when using accounts with restricted privileges.

Fix some types in the description for the recon script.

Fix a bug where the recon script never completed and kept processing the subnet
over and over due to an incorrect exit condition in the loop.

Change the packet count and timeout passed to ipmiping.  The previous values
required 9 seconds per address polled.  This meant it took almost 40 minutes
to scan one 256 address subnet.