Log plugin stderr and improve eventlog performance
The stderr output from plugins and executed programs is written to the agent log.

The way log events are filtered was split into two steps, this was done because getting the description is a relatively expensive operation so first we do the other checks to see if the log entry is interesting.  If so then the description is retrieved and the final filter check is done.