Fix CentOS / RHEL version detection
This change fixes three issues with version detection on RHEL distros.

First lsb_release -sd wraps the response in quotes.  This happens on
both CentOS 6.x and 7.x.

Second CentOS 7 (and probably RHEL 7) don't install the package with
lsb_release by default so this change gets the information from
/etc/redhat-release if lsb_release doesn't return anything.

Finally there are extra spaces at the end of the string.