Force omnishell_client to use JSON::PP.
The binary compiled with PAR::Packer for Pandora FMS doesn't work well
with JSON:XS, probably because JSON::Backend::XS is defined in the __DATA__
section of JSON.pm and that doesn't work well with PAR::Filter. If this
becomes a bottleneck, a workaround would be possible (e.g., redefining
JSON::Backend::XS in omnishell_client.pl).